Move WebArrayBufferConverter.h from to public/web/ so Chromium can use it
This file was added in r165330 as part of a three-sided patch to remove
v8 dependencies from the platform API, but was placed in a directory
inaccessible to the embedder. This patch simply fixes that location (and
moves some includes around to satisfy DEPS).

R=abarth@chromium.org
BUG=297477

Review URL: https://codereview.chromium.org/154423003

git-svn-id: svn://svn.chromium.org/blink/trunk@166514 bbb929c8-8fbe-4397-9dbb-9b2b20218538