Add an if(frame) check to toIsolate(Frame*)
- The crash happens in Youtube in a very weird path
(https://code.google.com/p/chromium/issues/detail?id=340567). To reproduce the
crash, we need to somehow call isURLAllowed() after the frame is gone.

- Actually the crash is a regression caused by
https://codereview.chromium.org/151283002/. Previously toIsolate(Frame*) was
implemented so that it accepts a null frame, but
https://codereview.chromium.org/151283002/ disallowed a null frame. This change
caused the crash.

- So this CL is going to remove all call sites that call toIsolate() with a null
frame.

For more details, see a crash log in the bug.

BUG=340567

Review URL: https://codereview.chromium.org/146083004

git-svn-id: svn://svn.chromium.org/blink/trunk@166562 bbb929c8-8fbe-4397-9dbb-9b2b20218538