Fixed a crash caused by an incorrect assumption in PerformanceTiming that a document loader would always be present.
This isn't the case when this code is triggered by sync script execution when a PPAPI plugin is shutdown, as the relevant frame has already been detached.

When this happens we just return timestamps of 0 instead, as we have no way of computing wall time.

R=japhet@chromium.org
BUG=341589

Review URL: https://codereview.chromium.org/142713008

git-svn-id: svn://svn.chromium.org/blink/trunk@166874 bbb929c8-8fbe-4397-9dbb-9b2b20218538