TimelineFlameChart: set minimum window size to 1000ms.
From the UI perspective flame chart looks better if it doesn't scale on the fly.
So we need to set 1000ms as a default time range for it.

Drive by fix: keep timelineData uninitialized in constructor.
It fixes the problem with empty flame charts if events were recorded in another mode.

BUG=316167

Review URL: https://codereview.chromium.org/159763003

git-svn-id: svn://svn.chromium.org/blink/trunk@166952 bbb929c8-8fbe-4397-9dbb-9b2b20218538