Remove isolated world parameters from event listener attributes
Currently, when accessing an event listener attribute,
we get the current isolated world in the binding layer,
pass it down to the core layer, and then use it to request
V8AbstractEventListener that belongs to the same isolated world.
However, we don't need to pass the current isolated world around in this way.
Alternately, we just need to get the current isolated world
when V8AbstractEventListener is requested.

This CL implements V8AbstractEventListener::belongToTheCurrentWorld() and
removes a bunch of world parameters from the code base.

Review URL: https://codereview.chromium.org/149803005

git-svn-id: svn://svn.chromium.org/blink/trunk@167001 bbb929c8-8fbe-4397-9dbb-9b2b20218538