WebFrame exports size information for remote frames.
WebFrameImpl::setIsRemote() becomes an initialization point for remote
frames, called when a child frame is swapped out. Right now it only
causes necessary rendering parameters to be exported and sent to the
new child frame renderer process, later on it can be extended to
trigger the RemoteFrame creation process.

The size information is necessary for the child frame renderer to know
the viewport size that it is rendering into. If the frame size changes
dynamically we currently don't have a way to propagate this. FrameView
seems like the best place to trap a size change event but the current
plan is to get rid of that for out of process iframes. We'll have to
add that when we have a better idea of how RemoteFrame rendering will
work.

R=eseidel@chromium.org
BUG=325803

Review URL: https://codereview.chromium.org/135163006

git-svn-id: svn://svn.chromium.org/blink/trunk@167201 bbb929c8-8fbe-4397-9dbb-9b2b20218538