Remove CSS regions support, keeping a bare minimum to support "region-based" multicol.
Remove everything strictly specific to CSS regions, but keep the parts that
"region-based" multicol needs. So the render flow thread framework is still around,
albeit heavily trimmed down. RenderRegion has also been kept, but is now obviously
a misnomer (to be fixed later). It serves as a base class for what css3-break refers
to as a "fragmentainer" (e.g. column, page or region). Not sure if it should really
be a renderer type, seeing that these need to be created during layout (which should
be avoided for renderers). A fragmentainer is more akin to what a line box represents
than to what a renderer does.

Additional removal and cleanup is possible, but let's start with this.

BUG=none

Review URL: https://codereview.chromium.org/134473008

git-svn-id: svn://svn.chromium.org/blink/trunk@167275 bbb929c8-8fbe-4397-9dbb-9b2b20218538