Reland r167107.
Provide THUMB version of pushAllRegisters trampoline when -mthumb is on.

Interworking does not seem to work correctly for component build and pushAllRegisters ends up being called from THUMB code without correct mode switch which causes SIGILL. When the whole Blink is built with -mthumb it is perfectly correct to build pushAllRegisters as thumb code as well (only r4 - r7 need to be saved).

This version relies on CPP directives for conditional compilation instead of using .if/.else.

BUG=

Review URL: https://codereview.chromium.org/169453004

git-svn-id: svn://svn.chromium.org/blink/trunk@167305 bbb929c8-8fbe-4397-9dbb-9b2b20218538