Fixes for allowing to run layout tests on Android again
Follow up to r167125, which broke running layout tests on Android, as
the port implementation overrides the start_http_server() implementation.

Also follows up to r167129, which removes the support for additional_dirs
from the ApacheHTTP implementation. Android uses this (together with the
forwarder) to avoid pushing all layout tests to the device, and instead
serve them from the host which is running the tests.

TBR=dpranke, scottmg

Review URL: https://codereview.chromium.org/170753002

git-svn-id: svn://svn.chromium.org/blink/trunk@167357 bbb929c8-8fbe-4397-9dbb-9b2b20218538